+++ engine/src/Network/NetworkManager.hpp
@@ -127,7 +127,7 @@ public:
       * @param name The string to register.
       * @returns The new Id.
       */
-    uint32_t RegisterEvent(const QString& name);
+    uint16_t RegisterEvent(const QString& name);
 
     /**
       * Checks whether a string is already registered.
@@ -141,21 +141,21 @@ public:
       * @param id The Id to be checked.
       * @returns true If the Id is already used, otherwise false.
       */
-    bool EventRegistered(uint32_t id);
+    bool EventRegistered(uint16_t id);
 
     /**
       * Returns the Id for a string.
       * @param string The string to find.
       * @returns The Id for the string.
       */
-    uint32_t GetEventId(const QString& string);
+    uint16_t GetEventId(const QString& string);
 
     /**
       * Returns the string for an Id.
       * @param id The Id to find.
       * @returns The string for the Id.
       */
-    const QString& GetEventString(uint32_t id);
+    const QString& GetEventString(uint16_t id);
 
     /**
       * Returns next Id for automatic name generation that is available.
@@ -178,8 +178,8 @@ private:
     std::vector<std::shared_ptr<NetworkEvent>> mNetworkEventPrototypes;    //!< The list of prototypes known to mankind :P
     sf::UdpSocket mSocket;                                      //!< The socket used for data transmissions over network.
 
-    uint32_t mLastEventId;                                      //!< The Id used to register the last string with.
-    std::map<uint32_t, QString> mEventIds;                      //!< The relation map between Ids/strings.
+    uint16_t mLastEventId;                                      //!< The Id used to register the last string with.
+    std::map<uint16_t, QString> mEventIds;                      //!< The relation map between Ids/strings.
 };
 
 }
